+++ src/ui/layout/navbar-header.svelte
@@ -2,7 +2,7 @@
   import { type Snippet } from "svelte";
   import ModalSection from "./modal-section.svelte";
   import CloseButton from "../molecules/close-button.svelte";
-  import BackButton from "../components/back-button.svelte";
+  import BackButton from "../molecules/back-button.svelte";
   import IconArrow from "../atoms/icons/icon-arrow.svelte";
   import { LocalizationKeys } from "../localization/supportedLanguages";
   import Localized from "../localization/localized.svelte";
